+++ packages/api-models/src/queryParameters/detection/isNodeLineageParameters.ts
@@ -1,4 +1,4 @@
-import { type ValidationFaultCollector } from "@phylopic/utils"
+import { isUUIDv4, type ValidationFaultCollector } from "@phylopic/utils"
 import { type NodeEmbedded } from "../../types/NodeWithEmbedded"
 import { NODE_EMBEDDED_PARAMETERS } from "../constants/NODE_EMBEDDED_PARAMETERS"
 import { NodeLineageParameters } from "../types/NodeLineageParameters"
@@ -6,4 +6,6 @@ import { isListParameters } from "./isListParameters"
 export const isNodeLineageParameters = (
     x: unknown,
     faultCollector?: ValidationFaultCollector,
-): x is NodeLineageParameters => isListParameters<NodeEmbedded>(NODE_EMBEDDED_PARAMETERS)(x, faultCollector)
+): x is NodeLineageParameters =>
+    isListParameters<NodeEmbedded>(NODE_EMBEDDED_PARAMETERS)(x, faultCollector) &&
+    isUUIDv4((x as NodeLineageParameters).uuid, faultCollector?.sub("uuid"))