+++ packages/api-models/src/queryParameters/types/NodeImageTagsParameters.ts
@@ -1,0 +1,5 @@
+import { UUID } from "@phylopic/utils"
+import { DataParameters } from "./DataParameters"
+export type NodeImageTagsParameters = DataParameters & {
+    uuid: UUID
+}