+++ apps/api/CHANGELOG.md
@@ -9,11 +9,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+-   New endpoints: `/imagetags`, : `/nodes/{uuid}/imagetags`.
+-   New property: `Image.tags`.
+-   New query parameter for image list: `filter_tags`.
+
 ### Changed
+
+-   Upgraded `@aws-sdk` dependencies.
 
 ### Deprecated
 
 ### Fixed
+
+-   The `build` query parameter was msissing in the Serverless configuration for several endpoints.
 
 ### Removed